+++ swagger-viewer/app/scripts/contentscript/presentation/App/fixtures/SwaggerJson_2_0.ts
