+++ swagger-viewer/app/scripts/contentscript/presentation/App/fixtures/SwaggerYaml_2_0.ts
