+++ swagger-viewer/app/scripts/contentscript/presentation/App/fixtures/SwaggerYaml_3_0.ts
