+++ swagger-viewer/app/scripts/contentscript/presentation/App/index.test.tsx
