+++ swagger-viewer/app/scripts/contentscript/presentation/App/index.tsx
