+++ swagger-viewer/app/scripts/contentscript/presentation/index.tsx
